feat: [DATALAD RUNCMD] rename env vars: s/ES_/ELASTICSEARCH_/g
You should update your config management to write the new variables and run the same replacement
command over your git repo using this framework.
Because prod_role_aggregator.yaml is copied on instanziation, this is
not a breaking change. It will only take effect for new projects.

Justification of this change: Both ECS and vector.dev docs only use the
long version. The ES acronym is not commonly known.

=== Do not change lines below ===
{
 "chain": [],
 "cmd": "git ls-files -z | xargs --null -I '()' find './()' -type f -print0 | xargs --null sed --in-place --regexp-extended 's/ES_/ELASTICSEARCH_/g;'",
 "exit": 0,
 "extra_inputs": [],
 "inputs": [],
 "outputs": [],
 "pwd": "."
}
^^^ Do not change lines above ^^^