fix: host.uptime from RFC 5424 was extracted as 1/100s should be in s
Basically RFC 5424 [1] defines sysUpTime "The time (in hundredths of a
second) since the network management portion of the system was last
re-initialized." but ECS [2] defines it in "Seconds the host has been
up."

The conversion was missing resulting in wrong uptime for Syslog RFC 5424
events.

We cannot do much about the different definition between "network
management portion" and "host has been up". Should be negligible in practice.

[1]: https://www.rfc-editor.org/rfc/rfc5424#section-7.3.2
[2]: https://www.elastic.co/guide/en/ecs/current/ecs-host.html#field-host-uptime