feat: use compact function instead of new `del(.path, compact. true)`
As part of migration to 0.25.2, I saw in the release notes that `del`
can now compact as well. I considered if we want this but the code has
over 50 calls to `del` so it should be easier and more efficient to
compact the root event shortly before the event leaves the aggregator.

The `compact` functions exists for ages. I have just overlooked it
before.

Ref: https://github.com/vectordotdev/vector/pull/5231